+++ src/backend/joanie/payment/api.py
@@ -1,11 +1,39 @@
 """
 API endpoints
 """
+import logging
+
+from django.db import transaction
+
 from rest_framework import mixins, permissions, viewsets
-from rest_framework.decorators import permission_classes
-from .models import CreditCard
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+
+from joanie.core.models import User
+
+from . import exceptions, get_payment_backend, serializers
+
+logger = logging.getLogger(__name__)
 
 
+@api_view(["POST"])
+@transaction.atomic()
+def webhook(request):
+    """
+    The webhook called by payment provider
+    when a payment has been created/updated/refunded...
+    """
+
+    payment_backend = get_payment_backend()
+    try:
+        payment_backend.handle_notification(request)
+    except exceptions.ParseNotificationFailed as error:
+        return Response(str(error), status=error.status_code)
+
+    return Response(status=200)
+
+
+# pylint: disable=too-many-ancestors
 class CreditCardViewSet(
     mixins.ListModelMixin,
     mixins.RetrieveModelMixin,